backend/requirements.txt
numpy>=1.26
scipy>=1.12
requests>=2.31
python-dotenv>=1.0
